Split DatabasesResourceTree and notebooksResourceTree component from ResourceTree
