+++ .eslintignore
@@ -192,3 +192,5 @@ src/Explorer/Notebook/temp/inputs/connected-editors/codemirror.tsx
 src/Explorer/Tree/ResourceTreeAdapter.tsx
 __mocks__/monaco-editor.ts
 src/Explorer/Tree/ResourceTree.tsx
+src/Explorer/Tree/DatabasesResourceTree.tsx
+src/Explorer/Tree/NotebooksResourceTree.tsx
